requirements_pdf_processing.txt
# Enhanced PDF to DOCX Conversion Dependencies
# ===========================================

# Core PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
pdf2image>=1.16.0

# Word document processing
python-docx>=0.8.11

# OCR capabilities (for scanned documents)
pytesseract>=0.3.10
Pillow>=9.5.0

# Text processing and NLP
nltk>=3.8.1
spacy>=3.5.0

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0

# Install spaCy model for better text processing
# Run: python -m spacy download en_core_web_sm

# Optional: For better OCR performance
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu 